% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSigs.R
\name{groupSigs}
\alias{groupSigs}
\title{Group the signatures by treatment ligands}
\usage{
groupSigs(siglist, sigmeta)
}
\arguments{
\item{siglist}{siglist of LigScape}

\item{sigmeta}{sigmeta of LigScape}
}
\value{
a list siggroups:
siggroups$sigLs: the vector of ligands in LigScape
siggroups$siggroups: the list of signatureid-vectors of each ligand
}
\description{
Group the signatures by treatment ligands
}
